Implement Chapter Get-Mapping and change "approximate duration in minutes" field to integer

+++ src/main/java/de/mayer/backendspringpostgres/adventure/model/Chapter.java
@@ -2,11 +2,10 @@ package de.mayer.backendspringpostgres.adventure.model;
 
 import java.util.List;
 
-public record Chapter(
-        String name,
-        String subheader,
-        Double approximateDurationInMinutes,
-        List<RecordInAChapter> records
+public record Chapter(String name,
+                      String subheader,
+                      Integer approximateDurationInMinutes,
+                      List<RecordInAChapter> records
 ) {
 
     public Chapter {
