+++ src/main/java/de/mayer/backendspringpostgres/graph/model/Chapter.java
@@ -1,4 +1,4 @@
 package de.mayer.backendspringpostgres.graph.model;
 
-public record Chapter(String name, Double approximateDurationInMinutes) {
+public record Chapter(String name, Integer approximateDurationInMinutes) {
 }
