+++ src/main/java/de/mayer/backendspringpostgres/graph/model/Path.java
@@ -3,11 +3,11 @@ package de.mayer.backendspringpostgres.graph.model;
 import java.util.LinkedList;
 import java.util.stream.Collectors;
 
-public record Path(LinkedList<Chapter> chapters, Double approximateDurationInMinutes) {
+public record Path(LinkedList<Chapter> chapters, Integer approximateDurationInMinutes) {
 
     @Override
     public String toString() {
-        return "%s (%.2f Minutes)".formatted(
+        return "%s (%d Minutes)".formatted(
                 chapters.stream().map(Chapter::name).collect(Collectors.joining(" -> ")),
                 approximateDurationInMinutes);
     }
